+++ src/main/java/dev/farneser/tasktracker/api/operations/queries/kanbancolumn/getById/GetKanbanColumnByIdQuery.java
@@ -1,7 +1,7 @@
 package dev.farneser.tasktracker.api.operations.queries.kanbancolumn.getById;
 
 import dev.farneser.tasktracker.api.mediator.Query;
-import dev.farneser.tasktracker.api.operations.view.KanbanColumnView;
+import dev.farneser.tasktracker.api.operations.views.KanbanColumnView;
 import lombok.AllArgsConstructor;
 import lombok.Data;
