+++ src/main/java/dev/farneser/tasktracker/api/operations/queries/user/getbyemail/GetUserByEmailQuery.java
@@ -1,7 +1,7 @@
 package dev.farneser.tasktracker.api.operations.queries.user.getbyemail;
 
 import dev.farneser.tasktracker.api.mediator.Query;
-import dev.farneser.tasktracker.api.operations.view.UserView;
+import dev.farneser.tasktracker.api.operations.views.UserView;
 import lombok.AllArgsConstructor;
 import lombok.Data;
