+++ src/main/java/dev/farneser/tasktracker/api/operations/queries/user/getbyemail/GetUserByEmailQueryHandler.java
@@ -3,7 +3,7 @@ package dev.farneser.tasktracker.api.operations.queries.user.getbyemail;
 import dev.farneser.tasktracker.api.exceptions.NotFoundException;
 import dev.farneser.tasktracker.api.exceptions.UserNotFoundException;
 import dev.farneser.tasktracker.api.mediator.QueryHandler;
-import dev.farneser.tasktracker.api.operations.view.UserView;
+import dev.farneser.tasktracker.api.operations.views.UserView;
 import dev.farneser.tasktracker.api.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.modelmapper.ModelMapper;
