+++ src/main/java/dev/farneser/tasktracker/api/service/BaseService.java
@@ -3,7 +3,7 @@ package dev.farneser.tasktracker.api.service;
 import dev.farneser.tasktracker.api.exceptions.NotFoundException;
 import dev.farneser.tasktracker.api.mediator.Mediator;
 import dev.farneser.tasktracker.api.operations.queries.user.getbyemail.GetUserByEmailQuery;
-import dev.farneser.tasktracker.api.operations.view.UserView;
+import dev.farneser.tasktracker.api.operations.views.UserView;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.modelmapper.ModelMapper;
