+++ src/main/java/dev/farneser/tasktracker/api/service/KanbanColumnService.java
@@ -4,7 +4,7 @@ import dev.farneser.tasktracker.api.exceptions.NotFoundException;
 import dev.farneser.tasktracker.api.mediator.Mediator;
 import dev.farneser.tasktracker.api.operations.commands.kanbancolumn.create.CreateKanbanColumnCommand;
 import dev.farneser.tasktracker.api.operations.queries.kanbancolumn.getById.GetKanbanColumnByIdQuery;
-import dev.farneser.tasktracker.api.operations.view.KanbanColumnView;
+import dev.farneser.tasktracker.api.operations.views.KanbanColumnView;
 import dev.farneser.tasktracker.api.web.dto.KanbanColumnDto;
 import lombok.extern.slf4j.Slf4j;
 import org.modelmapper.ModelMapper;
